% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeHeatmap.R
\name{TreeHeatmap}
\alias{TreeHeatmap}
\title{Heatmap at arbitary levels of a tree}
\usage{
TreeHeatmap(tree, tree_fig, hm_data, tree_hm_gap = 0, rel_width = 1,
  column_order = NULL, column_split = NULL, column_split_gap = 0.2,
  column_anno_color = NULL, column_anno_gap = 0.1,
  legend_title_hm = "Expression", legend_title_column_anno = "group",
  show_colnames = FALSE, colnames_position = "top",
  colnames_angle = 0, colnames_offset_x = 0, colnames_offset_y = 0,
  colnames_size = 4, colnames_hjust = 0.5, show_rownames = FALSE,
  rownames_angle = 0, rownames_offset_x = 0, rownames_offset_y = 0,
  rownames_size = 4, rownames_hjust = 0.5)
}
\arguments{
\item{tree}{A phylo object}

\item{tree_fig}{A ggtree object that outputs from
\code{\link[ggtree]{ggtree}}.}

\item{hm_data}{A data frame. Data to plot heatmap. Its rownames should be
able to match to nodes of the \strong{tree}.}

\item{tree_hm_gap}{A numeric value to specify the gap between the tree and
the heatmap.}

\item{rel_width}{A numeric value to specify the width of heatmap relative to
the width of the tree. For example, \code{rel_width = 1}, the width of
heatmap is the same as the width of the tree.}

\item{column_order}{A character vector that includes the column name of
\strong{hm_data} to specify the display order of the heatmap. It's ignored
when \strong{column_split} is provided.}

\item{column_split}{A named character vector that gives the group information
about columns to split the heatmap. It's named by the colnames of
\strong{hm_data}.}

\item{column_split_gap}{A numeric value to specify the gap between the
columns of heatmap that are split.}

\item{column_anno_color}{A named vector to specify colors that are used to
annotate columns of heatmap.}

\item{column_anno_gap}{A numeric value to specify the gap between the
column annotation bar and the heatmap.}

\item{legend_title_hm}{The legend title of the heatmap.}

\item{legend_title_column_anno}{The legend title of the column annotation.}

\item{show_colnames}{A logical value to specify whether column names should
be displayed. The default is FALSE.}

\item{colnames_position}{The position of column names, "top" or "bottom".}

\item{colnames_angle}{A numeric value. The angle of column names.}

\item{colnames_offset_x}{A numeric value to shift column names on x-axis. The
defaut is 0.}

\item{colnames_offset_y}{A numeric value to shift column names on y-axis. The
defaut is 0.}

\item{colnames_size}{A numeric value to specify the size of column names.}

\item{colnames_hjust}{The hjust for column names: 0 (left aligned); 0.5
(centered); 1 (right aligned).}

\item{show_rownames}{A logical value to specify whether row names should
be displayed. The default is FALSE.}

\item{rownames_angle}{A numeric value. The angle of row names.}

\item{rownames_offset_x}{A numeric value to shift row names on x-axis. The
defaut is 0.}

\item{rownames_offset_y}{A numeric value to shift row names on y-axis. The
defaut is 0.}

\item{rownames_size}{A numeric value to specify the size of row names.}

\item{rownames_hjust}{The hjust for row names: 0 (left aligned); 0.5
(centered); 1 (right aligned).}
}
\description{
\code{TreeHeatmap} displays a heatmap at a arbitary level of a tree.
}
\examples{

library(TreeSummarizedExperiment)
library(ggtree)
library(ggplot2)

data(tinyTree)

# Simulate data at different levels of a toy tree
set.seed(1)
p1 <- rep(0.1, 10)
p2 <- c(0.05, 0.05, 0.05, 0.175, 0.175, rep(0.1, 5))
ct <- cbind(rmultinom(n = 2, size = 100, prob = p1),
            rmultinom(n = 2, size = 100, prob = p2))
rownames(ct) <- transNode(tree = tinyTree, node = 1:10)

# leaf level: leaf nodes
lse <- TreeSummarizedExperiment(assays = list(ct),
                                rowTree = tinyTree)

# all levels: data on an internal node is generated by taking the sum
# of values from its descendant leaf nodes.
lev <- showNode(tree = tinyTree, only.leaf = FALSE)
tse <- aggValue(x = lse, rowLevel = 1:19, FUN = sum)
tse

# select an arbitary level
sel <- c(18, 6:9, 13, 10)
sse <- tse[sel, ]
hm_data <- assays(sse)[[1]]
rownames(hm_data) <- rowLinks(sse)$nodeLab
colnames(hm_data) <- paste0("S", 1:4)

# prepare the tree figure
tree_fig <- ggtree(tinyTree,
                   branch.length = "none",
                   layout = "rectangular") +
    #geom_text2(aes(label = label)) +
    geom_hilight(node = 18, fill = "orange", alpha = 0.3) +
    geom_hilight(node = 13, fill = "blue", alpha = 0.3)

fig <- TreeHeatmap(tree = tinyTree, tree_fig = tree_fig,
                   hm_data = data.frame(hm_data),
                   tree_hm_gap = 0.1,
                   rel_width = 2,
                   column_order = c("S4", "S3", "S2", "S1"),
                   column_split = c("S3" = "B", "S4" = "B",
                                    "S2" = "A", "S1" = "C"),
                   column_anno_color = c("S1" = "black",
                                         "S2" = "blue",
                                         "S3" = "orange",
                                         "S4" = "orange"),
                   column_split_gap = 0.3,
                   column_anno_gap = 0.1,
                   show_colnames = TRUE,
                   colnames_offset_y = 0.4,
                   colnames_angle = 0,
                   show_rownames = TRUE,
                   rownames_size = 3,
                   rownames_offset_x = 0.1,
                   rownames_hjust = 0)
fig
# use expand_limits to display the missing part of row names
fig + expand_limits(x = c(0, 22))
}
\author{
Ruizhu Huang
}
